compile fixes + doxygen
