+++ dialects/clos/pass/rw/clos_conv_prep.cpp
@@ -2,7 +2,7 @@
 
 #include "thorin/pass/fp/eta_exp.h"
 
-#include "dialects/clos/clos_conv.h"
+#include "dialects/clos/clos.h"
 
 namespace thorin::clos {
